Updated splitting for known holdout
Previously, the known holdout data (single samples of data) was being taken from the 80% split of the known data that would be used for training. Now, the single samples of known data are being taken from the 20% split of known data which now cumulatively makes up both the known data to build the threshold and the data used to test it.