runAnomalyDetectionTests Now Incorporates unknownDataSplit()
When building the threshold, the number of data samples in the known and unknown splits should be much closer.
If the number of known data samples is greater than the number of unknown data samples, all unknown data is used for building the threshold. If not, then the amount of unknown data samples used for building threshold will equal known data samples.